+++ MvvmCross/Binding/Droid/MvvmCross.Binding.Droid.csproj
@@ -116,7 +116,6 @@
     <Compile Include="Views\IMvxAdapterWithChangedEvent.cs" />
     <Compile Include="Views\IMvxLayoutInflaterHolder.cs" />
     <Compile Include="Target\MvxViewLongClickBinding.cs" />
-    <Compile Include="Views\MvxBaseListItemView.cs" />
     <Compile Include="BindingContext\MvxAndroidBindingContext.cs" />
     <Compile Include="Views\MvxContextWrapper.cs" />
     <Compile Include="Views\MvxExpandableListAdapter.cs" />
